Update with minimal example

+++ src/Console.cpp
@@ -319,7 +319,7 @@ size_t Console::write(const uint8_t *buf, size_t size) {
 #define PREFIX_LEN (15)
 void Console::debugPrefix(char* s) {
   uint32_t t = ::millis();
-  sprintf(s, "%6d.%03d: ",t/1000,t%1000);
+  sprintf(s, "%6d.%03d: ",(int)t/1000,(int)t%1000);
 }
 
 void Console::debugf(const char* format, ...) {
